+++ modules/overrides/config_test.go
@@ -107,6 +107,8 @@ metrics_generator_collection_interval: 5s
 metrics_generator_disable_collection: false
 metrics_generator_forwarder_queue_size: 6
 metrics_generator_forwarder_workers: 7
+metrics_generator_remote_write_headers:
+  tenant-id: foo
 metrics_generator_processor_service_graphs_histogram_buckets: [1,2]
 metrics_generator_processor_service_graphs_dimensions: ['foo']
 metrics_generator_processor_service_graphs_peer_attributes: ['foo']
@@ -171,6 +173,8 @@ defaults:
     max_active_series: 4
     collection_interval: 5s
     disable_collection: false
+    remote_write_headers:
+      tenant-id: foo
     forwarder:
       queue_size: 6
       workers: 7